removed queue and tasks
