+++ go.mod
@@ -6,5 +6,4 @@ require (
 	github.com/go-telegram/bot v0.8.2
 	github.com/google/uuid v1.4.0
 	github.com/joho/godotenv v1.5.1
-	golang.org/x/exp v0.0.0-20231110203233-9a3e6036ecaa
 )
