+++ justfile
@@ -1,2 +1,4 @@
 build:
 	go build -o build/botatobot cmd/botatobot/main.go
+run: build
+	./build/botatobot
