+++ pkg/commands.go
@@ -5,5 +5,4 @@ type Command string
 const (
 	HelpCmd     Command = "/help"
 	GenerateCmd Command = "/generate"
-	StatusCmd   Command = "/status"
 )
