+++ pkg/handlers.go
@@ -6,170 +6,82 @@ import (
 	"fmt"
 	"log"
 	"path/filepath"
+	"strings"
 
 	"github.com/go-telegram/bot"
 	"github.com/go-telegram/bot/models"
 	"github.com/google/uuid"
 )
 
-func getMessageOrUpdate(update *models.Update) (*models.Message, error) {
+func GenerateHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
 
 	if update == nil {
-		return &models.Message{}, fmt.Errorf("empty update")
+		log.Println("got an empty update, skipping")
+		return
 	}
 
 	message := update.Message
 
-	if message != nil {
-		return message, nil
+	if message == nil {
+		log.Printf("got no message from non-empty update, skipping")
+		return
 	}
 
-	edited := update.EditedMessage
+	// prompt is message.Text after removing the command
+	prompt := message.Text[(len(GenerateCmd) + 1):]
+	prompt = strings.TrimSpace(prompt)
 
-	if edited != nil {
-		return edited, nil
-	}
+	// TODO: sanitize prompt
 
-	return &models.Message{}, fmt.Errorf("no message found in update")
-}
+	log.Printf("User `%s` requested `%s`\n", message.Chat.Username, prompt)
 
-type Outcome interface {
-	GetMessage() *models.Message
-	GetIdentifier() uuid.UUID
-	GetDescription() string
-	Result() ([]byte, error)
-}
+	output, err := GenerateImage(prompt)
 
-func SendOutcome(ctx context.Context, b *bot.Bot) func(req Outcome) error {
-	return func(req Outcome) error {
+	if err != nil {
 
-		message := req.GetMessage()
-
-		res, err := req.Result()
-
-		if err != nil {
-
-			_, err := b.SendMessage(ctx, &bot.SendMessageParams{
-				ChatID:           message.Chat.ID,
-				Text:             fmt.Sprintf("Sorry, but something went wrong 😭 %s", err),
-				ReplyToMessageID: message.ID,
-			})
-
-			return err
-		}
-
-		_, err = b.SendPhoto(ctx, &bot.SendPhotoParams{
-			ChatID:  message.Chat.ID,
-			Caption: fmt.Sprint(req.GetDescription()),
-			Photo: &models.InputFileUpload{
-				Data:     bytes.NewReader(res),
-				Filename: filepath.Base(fmt.Sprintf("%s.png", req.GetIdentifier())),
-			},
-			DisableNotification: true,
+		_, err := b.SendMessage(ctx, &bot.SendMessageParams{
+			ChatID:           message.Chat.ID,
+			Text:             fmt.Sprintf("Sorry, but something went wrong 😭 %s", err),
+			ReplyToMessageID: message.ID,
 		})
 
-		return err
-	}
-}
-
-func Status(q Queue) bot.HandlerFunc {
-
-	return func(ctx context.Context, b *bot.Bot, update *models.Update) {
-
-		if update == nil {
-			log.Println("empty update")
-			return
-		}
-
-		message := update.Message
-
-		if message == nil {
-			log.Println("empty message")
-			return
-		}
-
-		isWorking := q.IsWorking()
-		numJobs := q.Len()
-
-		if !isWorking && numJobs == 0 {
-			b.SendMessage(ctx, &bot.SendMessageParams{
-				ChatID: message.Chat.ID,
-				Text:   "I am doing nothing and there are no jobs in the queue 🤖",
-			})
-			return
-		}
-
-		if isWorking {
-
-			b.SendMessage(ctx, &bot.SendMessageParams{
-				ChatID: message.Chat.ID,
-				Text:   fmt.Sprintf("I am generating an image and the queue has %d more jobs", numJobs),
-			})
-			return
-		}
-
-		if numJobs > 0 {
-			b.SendMessage(ctx, &bot.SendMessageParams{
-				ChatID: message.Chat.ID,
-				Text:   fmt.Sprintf("I am doing nothing and the queue has %d more jobs. That's weird!! ", numJobs),
-			})
-			return
-		}
-	}
-}
-
-func Generate(q Queue) bot.HandlerFunc {
-
-	return func(ctx context.Context, b *bot.Bot, update *models.Update) {
-
-		if update == nil {
-			log.Println("empty update")
-			return
-		}
-
-		message, err := getMessageOrUpdate(update)
-
 		if err != nil {
-			log.Printf("Failed to get message: %s", err)
-			return
+			log.Printf("Error sending message: %s", err)
 		}
-
-		err = q.Push(message)
-
-		log.Printf("Requested %s\n", message.Text)
-
-		if err != nil {
-			b.SendMessage(ctx, &bot.SendMessageParams{
-				ChatID:           message.Chat.ID,
-				Text:             fmt.Sprintf("Sorry, but your request was rejected 😬 %s", err),
-				ReplyToMessageID: message.ID,
-			})
-			log.Printf("Requested %s but rejected by %s\n", message.Text, err)
-			return
-		}
+		return
 	}
 
+	_, err = b.SendPhoto(ctx, &bot.SendPhotoParams{
+		ChatID:  message.Chat.ID,
+		Caption: fmt.Sprint(prompt),
+		Photo: &models.InputFileUpload{
+			Data:     bytes.NewReader(output),
+			Filename: filepath.Base(fmt.Sprintf("%s.png", uuid.New())),
+		},
+		DisableNotification: true,
+	})
+
+	if err != nil {
+		log.Printf("Error sending photo: %s", err)
+	}
 }
 
-func Help() bot.HandlerFunc {
+func HelpHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
 
-	return func(ctx context.Context, b *bot.Bot, update *models.Update) {
-
-		if update == nil {
-			log.Println("empty update")
-			return
-		}
-
-		message := update.Message
-
-		if message == nil {
-			log.Println("empty message")
-			return
-		}
-
-		b.SendMessage(ctx, &bot.SendMessageParams{
-			ChatID: message.Chat.ID,
-			Text:   "Hi! I'm a 🤖 that generates images from text. Use the /generate command follow by a prompt, like this: \n\n   /generate a cat in space \n\nBy default I will generate 5 images, but you can modify the seed, guidance and steps like so\n\n /generate a cat in space &seed_1234 &steps_50 &guidance_7.5\n\nCheck my status with /status\n\nHave fun!",
-		})
+	if update == nil {
+		log.Println("empty update")
+		return
 	}
+
+	message := update.Message
+
+	if message == nil {
+		log.Println("empty message")
+		return
+	}
+
+	b.SendMessage(ctx, &bot.SendMessageParams{
+		ChatID: message.Chat.ID,
+		Text:   "Hi! I'm a 🤖 that generates images from text. Use the /generate command follow by a prompt, like this: \n\n   /generate a cat in space \n\nBy default I will generate 5 images, but you can modify the seed, guidance and steps like so\n\n /generate a cat in space &seed_1234 &steps_50 &guidance_7.5\n\nCheck my status with /status\n\nHave fun!",
+	})
 }
